+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Version 1.1.1-SNAPSHOT:
   - assertion control
   - attributes by object class
   - don't use copy control
+  - matched values control
   - modify-increment operation
   - password modify operation
   - read entry controls